work around for deletet address and dependencies r

+++ src/stores/databaseStore.js
@@ -2,13 +2,14 @@ import {useState, useCallback, useEffect} from 'react';
 import * as database from '../database/databaseHandler';
 import {useBetween} from 'use-between';
 import useCalender from './calenderStore';
+import {deleteDrivenRouteByRouteId} from '../database/database';
 
 export default function useDatabase() {
   const useShareCalender = () => useBetween(useCalender);
   const {selectedDate} = useShareCalender();
 
   const [addresses, setAddresses] = useState([]);
-  const [routes, setRoutes] = useState(undefined);
+  const [routes, setRoutes] = useState([]);
   const [drivenRoutes, setDrivenRoutes] = useState([]);
   const [drivenRoutesByDate, setDrivenRoutesByDate] = useState([]);
   const [routesAreHidden, setRoutesAreHidden] = useState(false);
@@ -107,7 +108,19 @@ export default function useDatabase() {
   };
 
   const deleteAddress = add_id => {
-    database.deleteAddress(add_id);
+    console.log('daStore, deleteAddress', add_id);
+    console.log(routes);
+    if (routes.length > 0) {
+      let deleteRouteList = routes.filter(
+        route => route.startAdd_id === add_id || route.destAdd_id === add_id,
+      );
+      console.log('deleteRouteList ', deleteRouteList);
+      deleteRouteList.forEach(route =>
+        deleteDrivenRouteByRouteId(route.route_id),
+      );
+      deleteRouteList.forEach(route => deleteRoute(route.route_id));
+      database.deleteAddress(add_id);
+    }
     loadAddressesCallback();
   };
